Hiring of workers. Also shuffled them around so they belong to the City now, makes more sense. In addition, I got rid of some warnings.

+++ src/types.h
@@ -58,7 +58,7 @@ struct V2 {
 };
 
 inline V2 v2(Coord coord) {
-	return {coord.x, coord.y};
+	return {(real32)coord.x, (real32)coord.y};
 }
 
 typedef SDL_Color Color;
